feat(main): added sort type product in the main page
