+++ client/src/pages/main.js
@@ -3,148 +3,62 @@ import { Component } from '/react/newVersion/Component'
 import { dispatch } from '../redux/redux-store'
 import { addProducts } from '../redux/main-reducer'
 import UsersAPI from '../api/UsersAPI'
-import ProductCard from '../components/Main/ProductCard'
-import { getDynamicProducts } from '../redux/main-selectors'
+import ContainerProducts from '../components/Main/ContainerProducts'
 import { getAdditionally } from '../redux/additionally-reducer'
-import { getAdditionallyAll } from '../redux/additionally-selectors'
-import { getWishlist } from '../redux/wishlist-selectors'
+import { getAllProducts } from '../redux/main-selectors'
 
 class Main extends Component {
   constructor(props) {
     super(props)
     this.state = {
-      search: router.getSearch(),
       isReady: false,
-      totalCount: 0,
-      countProductsCard: 12,
-      dynamicListProducts: [],
-      items: [],
     }
   }
 
   getProductsFetch() {
     UsersAPI.getAllProduct().then(
       ({ products, items, tiers, types, nations, currentCurs }) => {
-        // console.log('main products', products, items)
-        // this.getItemsFetch ()
         dispatch(addProducts({ products, items }))
         dispatch(
           getAdditionally({ tiers, typesVichels: types, nations, currentCurs })
         )
+
         this.setState({
           ...this.state,
           isReady: true,
-          totalCount: products.length,
         })
-        this.dynamicAddProducts()
       }
     )
   }
 
-  getItemsFetch() {
-    UsersAPI.getAllItems().then((items) => {
-      this.setState({
-        ...this.state,
-        items: items.slice(0),
-      })
-    })
-  }
-
-  updateSearch() {
-    if (this.state.search !== router.search) {
-      this.setState({ ...this.state, search: router.search })
-      this.getProductsFetch()
-      //this.getItemInProduct()
-    }
-  }
-
-  scrollHandler = ({ target }) => {
-    const { scrollHeight, scrollTop } = target.documentElement
-    // debugger
-    if (
-      scrollHeight - (scrollTop + window.innerHeight) < 100 &&
-      this.state.dynamicListProducts.length < this.state.totalCount
-    ) {
-      this.dynamicAddProducts()
-      // console.log('fething')
-    }
-  }
-
-  dynamicAddProducts() {
-    const resultDynamic = getDynamicProducts(
-      this.state.countProductsCard,
-      this.state.dynamicListProducts.length
-    )
-
-    if (resultDynamic) {
-      this.setState({
-        ...this.state,
-        dynamicListProducts: [
-          ...this.state.dynamicListProducts,
-          ...resultDynamic,
-        ],
-      })
-    }
-  }
+  // shouldComponentUpdate(prevProps, prevState) {
+  //   if (prevProps !== this.props && prevState !== this.state) {
+  //     return true
+  //   }
+  // }
 
   componentDidMount() {
-    // subscriber.push(this.updateSearch.bind(this))
     this.getProductsFetch()
-    document.addEventListener('scroll', this.scrollHandler)
-    // window.addEventListener('resize', () => console.log('resize'))
-  }
-
-  componentDidUpdate(_, prevState) {
-    // debugger
-    if (prevState.dynamicListProducts !== this.state.dynamicListProducts) {
-      const grid = this._currentElement.dom
-
-      if (window.innerHeight > grid.clientHeight - 100) {
-        this.dynamicAddProducts()
-      }
-    }
-  }
-
-  componentWillUnmount() {
-    document.removeEventListener('scroll', this.scrollHandler)
-  }
-
-  addToCart = (id) => {
-    let cartItems = JSON.parse(localStorage.getItem('cart')) || []
-    if (cartItems.includes(id)) {
-      let index = cartItems.indexOf(id)
-      cartItems.splice(index, 1)
-    } else {
-      cartItems.push(id)
-    }
-    localStorage.setItem('cart', JSON.stringify(cartItems))
-    this.setState({ ...this.state })
-  }
-
-  addToWishlist = (id) => {
-    let cartItems = JSON.parse(localStorage.getItem('wishlist')) || []
-    if (cartItems.includes(id)) {
-      let index = cartItems.indexOf(id)
-      cartItems.splice(index, 1)
-    } else {
-      cartItems.push(id)
-    }
-    localStorage.setItem('wishlist', JSON.stringify(cartItems))
-    this.setState({ ...this.state })
   }
 
   render() {
+    const query = router.getQuery()
+    let products = getAllProducts()
+
+    if (this.state.isReady && query.type) {
+      // debugger
+      products = products.filter((product) =>
+        product.category.includes(query.type)
+      )
+    }
+
     return (
       <div className="grid">
-        {this.state.isReady
-          ? this.state.dynamicListProducts.map((product) => (
-              <ProductCard
-                card={product}
-                additionally={getAdditionallyAll()}
-                wishlist={getWishlist()}
-              />
-            ))
-          : 'Loading...'}
+        {this.state.isReady ? (
+          <ContainerProducts products={products} />
+        ) : (
+          'Loading...'
+        )}
       </div>
     )
   }
