+++ client/src/redux/main-reducer.js
@@ -1,16 +1,12 @@
-const UPDATE_CARRENT = 'UPDATE_CARRENT'
 const ADD_PRODUCTS = 'ADD_PRODUCTS'
 
 const initialState = {
-  carrent: 0,
   products: [],
   typeProductsAll: {},
 }
 
 const mainReducer = (state = initialState, action = {}) => {
   switch (action.type) {
-    case UPDATE_CARRENT:
-      return { ...state, ...action.body }
     case ADD_PRODUCTS: {
       return {
         ...state,
@@ -23,11 +19,6 @@ const mainReducer = (state = initialState, action = {}) => {
       return state
   }
 }
-
-export const updateCarrent = (carrent) => ({
-  type: UPDATE_CARRENT,
-  body: { carrent },
-})
 
 export const addProducts = ({ products, items }) => {
   const typeProductsAll = {}
